refactor(users): consolidate module, secure routes, and enforce JWT sub
- Controller: protect POST/GET/PATCH with AuthGuard('jwt') and @ApiBearerAuth()

- Controller: inject subject via @AuthUser('sub') instead of @Req()

- CreateUserDto: mark auth0Id as optional + readOnly; keep mapping for Auth0 fields

UpdateUserDto: extend UserDto with partialType and omit auth0Id and email

- Service: set auth0Id from JWT sub

- Service: update() now uses { new: true, runValidators: true }
