+++ API/src/modules/auth/auth.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { AuthService } from './auth.service';
 import { JwtStrategy } from './jwt.strategy';
-import { UsersModule } from '../users/users.module';
+import { UserModule } from '../user/user.module';
 
 @Module({
-  imports: [PassportModule.register({ defaultStrategy: 'jwt' }), UsersModule],
+  imports: [PassportModule.register({ defaultStrategy: 'jwt' }), UserModule],
   providers: [JwtStrategy, AuthService],
   // providers: [JwtStrategy],
   exports: [PassportModule],
