+++ API/src/modules/user/user.module.ts
@@ -4,11 +4,6 @@ import { MongooseModule } from '@nestjs/mongoose';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
 
-// import { MongooseModule } from '@nestjs/mongoose';
-// import { UserService } from './service/user.service';
-// import { UserController } from './controller/user.controller';
-// import { User, UserSchema } from './schema/user.schema';
-
 @Module({
   imports: [
     MongooseModule.forFeature([{ name: User.name, schema: UserSchema }]),
