+++ API/src/modules/user/user.service.ts
@@ -1,26 +1,81 @@
-import { Injectable } from '@nestjs/common';
-import { InjectModel } from '@nestjs/mongoose';
-import { User } from './schema/user.schema';
 import { Model } from 'mongoose';
+import { InjectModel } from '@nestjs/mongoose';
+import {
+  BadRequestException,
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
+import { User } from './schema/user.schema';
+import { CreateUserDto } from './dto/create-user.dto';
+import { UpdateUserDto } from './dto/update-user.dto';
 
 @Injectable()
 export class UserService {
-  constructor(@InjectModel(User.name) private userModel: Model<User>) {}
+  constructor(
+    @InjectModel(User.name) private readonly userModel: Model<User>,
+  ) {}
 
-  async findUserById(id: string): Promise<User | null> {
-    return this.userModel.findById(id).exec();
+  // CREATE USER
+  async create(sub: string, createUserDto: CreateUserDto): Promise<User> {
+    if (!sub) {
+      throw new BadRequestException('A valid auth0 sub is required.');
+    }
+    // take auth0Id (sub) from Access token (passed in via controller) rather than request body/ dto
+    const userProfile = { ...createUserDto, auth0Id: sub };
+    try {
+      return await this.userModel.create(userProfile);
+    } catch (error) {
+      console.log('Error creating user.', error);
+      throw new InternalServerErrorException('Error creating user.');
+    }
   }
 
-  async createUser(userData: Partial<User>): Promise<User> {
-    const newUser = new this.userModel(userData);
-    return newUser.save();
+  // FIND USER BY AUTH) ID
+  async findOneByAuth0Id(sub: string): Promise<User> {
+    try {
+      const user = await this.userModel.findOne({ auth0Id: sub }).exec();
+      if (!user) {
+        throw new NotFoundException('User not found.');
+      }
+      return user;
+    } catch (error) {
+      console.log('Error finding user.', error);
+      throw new InternalServerErrorException(
+        'Internal server error while fetching user.',
+      );
+    }
   }
 
-  async updateUser(id: string, userData: Partial<User>): Promise<User | null> {
-    return this.userModel.findByIdAndUpdate(id, userData, { new: true }).exec();
+  // UPDATE USER
+  async update(sub: string, updateUserDto: UpdateUserDto): Promise<User> {
+    if (!sub) {
+      throw new BadRequestException('A valid auth0 sub is required.');
+    }
+    try {
+      const updatedUser = await this.userModel
+        .findOneAndUpdate({ auth0Id: sub }, updateUserDto, {
+          new: true,
+          runValidators: true,
+        })
+        .exec();
+      if (!updatedUser) {
+        throw new NotFoundException('Failed to update user details.');
+      }
+      return updatedUser;
+    } catch (error) {
+      console.log('Error updating user.', error);
+      throw new InternalServerErrorException(
+        'Internal server error while updating user.',
+      );
+    }
   }
 
-  async deleteUser(id: string): Promise<User | null> {
-    return this.userModel.findByIdAndDelete(id).exec();
-  }
+  // async deleteUser(id: string): Promise<User | null> {
+  //   return this.userModel.findByIdAndDelete(id).exec();
+  // }
 }
+
+// // getUserByauth0Sub = async (auth0Sub) => {}
+// // createUser = async (auth0Sub, userProfile) => {}
+// // updateUserProfile = async (auth0Sub, updates) => {}
